requirements.txt
dvc==2.35.2
dvc_gdrive==2.19.1
pandas==1.5.2
PyYAML==6.0
scikit-surprise
matplotlib
jupyter
torch
torch_geometric
torch_sparse
torch_scatter
sklearn
numpy
pandas_profiling
ray
tqdm
argparse
recmetrics
DGL
gnnlens